Time: 84 ms (90.54%), Space: 32.4 MB (85.11%) - LeetHub

+++ 1976-number-of-ways-to-arrive-at-destination/1976-number-of-ways-to-arrive-at-destination.cpp
@@ -24,7 +24,7 @@ public:
             
             for(auto c:adj[node]){
                 int adjnode=c.first;
-                long long wt=c.second;
+                int wt=c.second;
                 if(dist+wt<dis[adjnode]){
                     dis[adjnode]=dist+wt;
                     pq.push({dis[adjnode],adjnode});
